@@ -1,45 +1,63 @@
 source 'https://rubygems.org'
+git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-ruby '2.4.0'
+ruby '2.4.4'
+
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-#gem 'rails', '4.2.5.1'
-gem 'rails', '>= 4.2.5.1'
-# Bootstrap
-gem 'bootstrap-sass', '>= 3.2.0.0'
+gem 'rails', '~> 5.2.1'
+
+# Use Puma as the app server
+gem 'puma', '~> 3.7'
+
+# Use Paperclip to saving images in FS
+gem 'paperclip', '~> 5.0.0'
+
+# Postgres
+gem 'pg'
+gem 'rails_12factor'
+
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 5.0'
-# Use Uglifier as compressor for JavaScript assets
-gem 'uglifier', '>= 1.3.0'
-# Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
-# See https://github.com/rails/execjs#readme for more supported runtimes
-# gem 'therubyracer', platforms: :ruby
+gem 'sass-rails', '>= 5.0.7'
+gem 'bootstrap-sass'
+
 # Use jquery as the JavaScript library
-gem 'jquery-rails'
-# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
-gem 'turbolinks'
+gem 'jquery-rails', '~> 4.3'
+
+# Use CoffeeScript for .coffee assets and views
+gem 'coffee-rails', '~> 4.2'
+
+# Turbolinks makes navigating your web application faster. Read more: https://github.com/turbolinks/turbolinks
+gem 'turbolinks', '~> 5'
+
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
+gem 'jbuilder', '~> 2.5'
+
+# Windows does not include zoneinfo files, so bundle the tzinfo-data gem
+gem 'tzinfo-data', platforms: [:mingw, :mswin, :x64_mingw, :jruby]
+
+# React
+gem 'react-rails', '~> 1.7'
+
+# Manipulate images with minimal use of memory via ImageMagick / GraphicsMagick (https://github.com/minimagick/minimagick)
+gem "mini_magick"
+
+# Ruby binding to ImageMagick (https://github.com/rmagick/rmagick)
+gem 'rmagick', '~> 2.16.0', require: 'rmagick'
+
+# Embed the V8 JavaScript interpreter into Ruby (http://github.com/cowboyd/therubyracer)
+gem 'therubyracer'
+
+# Simple, efficient background processing for Ruby (http://sidekiq.org)
+gem 'sidekiq'
+
 # bundle exec rake doc:rails generates the API under doc/api.
 gem 'sdoc', '~> 0.4.0', group: :doc
 
 # Use ActiveModel has_secure_password
 gem 'bcrypt', '~> 3.1.7'
 
-# Use Paperclip to saving images in FS
-gem 'paperclip', '~> 5.0.0'
-
-# Use Unicorn as the app server
-# gem 'unicorn'
-
-# Use Capistrano for deployment
-# gem 'capistrano-rails', group: :development
-
 # For pagination
 gem 'kaminari'
-
-# For adding removing dynamicly
-gem 'cocoon', '~> 1.2.9'
 
 # ReCaptcha
 gem 'recaptcha', :require => 'recaptcha/rails'
