@@ -1,29 +1,18 @@
 # Load the Rails application.
-require File.expand_path('../application', __FILE__)
+require_relative 'application'
 
 # Initialize the Rails application.
 Rails.application.initialize!
 
 # Configure mail
 ActionMailer::Base.raise_delivery_errors = true
-if Rails.env.development?
+unless Rails.env.development?
   ActionMailer::Base.smtp_settings = {
-      :address        => 'smtp.gmail.com',
-      :domain         => 'mail.google.com',
-      :port           => 587,
-      :user_name      => 'derpyaitscore@gmail.com',
-      :password       => 'animespirit',
-      :authentication => :plain,
-      :enable_starttls_auto => true
-  }
-else
-  ActionMailer::Base.smtp_settings = {
-      :address        => 'smtp.gmail.com',
-      :domain         => 'mail.google.com',
-      :port           => 587,
-      :user_name      => ENV['MAIL_USER'],
-      :password       => ENV['MAIL_PASSWORD'],
-      :authentication => :plain,
-      :enable_starttls_auto => true
+      address: "smtp.sendgrid.net",
+      port: 25,
+      domain: "aitscore.com",
+      authentication: :plain,
+      user_name: ENV['SENDGRID_USERNAME'],
+      password: ENV['SENDGRID_PASSWORD']
   }
 end
