@@ -1,4 +1,4 @@
-class CreateInterests < ActiveRecord::Migration
+class CreateInterests < ActiveRecord::Migration[4.2]
   def change
     create_table :interests do |t|
       t.string :name
