@@ -1,4 +1,4 @@
-class CreateSchoolings < ActiveRecord::Migration
+class CreateSchoolings < ActiveRecord::Migration[4.2]
   def change
     create_table :schoolings do |t|
       t.string :name
