@@ -1,4 +1,4 @@
-class CreateInfos < ActiveRecord::Migration
+class CreateInfos < ActiveRecord::Migration[4.2]
   def change
     create_table :infos do |t|
       t.string :name
