@@ -1,4 +1,4 @@
-class CreateTutors < ActiveRecord::Migration
+class CreateTutors < ActiveRecord::Migration[4.2]
   def change
     create_table :tutors do |t|
       t.belongs_to :info, index: true, unique: true, foreign_key: true
