@@ -1,4 +1,4 @@
-class CreateStudents < ActiveRecord::Migration
+class CreateStudents < ActiveRecord::Migration[4.2]
   def change
     create_table :students do |t|
       t.string :code_name, unique: true
