@@ -1,4 +1,4 @@
-class CreatePictures < ActiveRecord::Migration
+class CreatePictures < ActiveRecord::Migration[4.2]
   def change
     create_table :pictures do |t|
       t.string :path
