@@ -1,4 +1,4 @@
-class CreatePictureInterests < ActiveRecord::Migration
+class CreatePictureInterests < ActiveRecord::Migration[4.2]
   def change
     create_table :picture_interests do |t|
       t.integer :earned_points #weight of link between interest and picture
