@@ -14,9 +14,15 @@
 //= require jquery_ujs
 //= require turbolinks
 //= require bootstrap
-//= require cocoon
 //= require Chart.bundle
 //= require chartkick
+//= require react
+//= require react_ujs
+//= require activestorage
+//= require components
+//= require lodash
+//= require heatmap.js/heatmap.min
+//= require html2canvas/html2canvas.min
 
 //Converts all text in class "date_to_local" to local time
 $(document).on('turbolinks:load', function () {
