@@ -1,5 +1,6 @@
 class AitscoreMailer < ApplicationMailer
   default from: 'derpy@aitscore.com'
+
   def registration_confirmation(user)
     send_mail(user, 'Registration Confirmation')
   end
@@ -9,8 +10,8 @@ class AitscoreMailer < ApplicationMailer
   end
 
   private
-  def send_mail(user, subject)
-    @user = user
-    mail(to: "#{user.last_name} #{user.name} <#{user.mail}>", subject: subject)
-  end
+    def send_mail(user, subject)
+      @user = user
+      mail(to: "#{user.last_name} #{user.name} <#{user.mail}>", subject: subject)
+    end
 end
