@@ -3,11 +3,13 @@
 # Version of your assets, change this if you want to expire all your assets.
 Rails.application.config.assets.version = '1.0'
 
-# Add additional assets to the asset load path
+# Add additional assets to the asset load path.
 # Rails.application.config.assets.paths << Emoji.images_path
+# Add Yarn node_modules folder to the asset load path.
+Rails.application.config.assets.paths << Rails.root.join('node_modules')
 
 # Precompile additional assets.
 # application.js, application.css, and all non-JS/CSS in app/assets folder are already added.
 # Rails.application.config.assets.precompile += %w( search.js )
-Rails.application.config.assets.precompile += %w( tests.css exiting_page_script.js tests.js students.js questions.js)
-#exiting_page_script.js
+Rails.application.config.assets.precompile += %w(tests.css exiting_page_script.js tests.js students.js questions.js)
+Rails.application.config.assets.precompile += %w[webgazer/webgazer.js heatmap.js/heatmap.min.js]
