@@ -1,9 +1,9 @@
 Recaptcha.configure do |config|
-  if Rails.env.development?
-    config.site_key = '6LeIxAcTAAAAAJcZVRqyHh71UMIEGNQ_MXjiZKhI'
-    config.secret_key = '6LeIxAcTAAAAAGG-vFI1TnRWxMZNFuojJ4WifJWe'
-  else
+  if Rails.env.production?
     config.site_key = ENV['RECAPTCHA_SITE_KEY']
     config.secret_key = ENV['RECAPTCHA_SECRET_KEY']
+  else
+    config.site_key = '6LeIxAcTAAAAAJcZVRqyHh71UMIEGNQ_MXjiZKhI'
+    config.secret_key = '6LeIxAcTAAAAAGG-vFI1TnRWxMZNFuojJ4WifJWe'
   end
 end
